Deprecate usage of XML Spring schemas
Move CustomAnalyzers12 test to annotation

Related to #141.
