+++ src/test/java/fr/pilato/spring/elasticsearch/it/annotation/rest/customanalyzers/AppConfig.java
@@ -1,0 +1,38 @@
+/*
+ * Licensed to David Pilato (the "Author") under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. Author licenses this
+ * file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package fr.pilato.spring.elasticsearch.it.annotation.rest.customanalyzers;
+
+import fr.pilato.spring.elasticsearch.ElasticsearchRestClientFactoryBean;
+import fr.pilato.spring.elasticsearch.it.annotation.rest.RestAppConfig;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class AppConfig extends RestAppConfig {
+
+	@Override
+	protected void enrichFactory(ElasticsearchRestClientFactoryBean factory) {
+		factory.setClasspathRoot("/models/root/custom-analyzers-12/client");
+		factory.setMappings(new String[] {"twitter/_doc"});
+		factory.setMergeSettings(true);
+		factory.setMergeMapping(true);
+		factory.setForceMapping(true);
+	}
+
+}
